Revisions to grid storage

+++ contrib/distributed/src/main/java/sim/app/dflockers/DFlockers.java
@@ -40,8 +40,7 @@ public class DFlockers extends DSimState {
 	public DFlockers(final long seed) {
 		super(seed, DFlockers.width, DFlockers.height, DFlockers.neighborhood);
 
-		//final double[] discretizations = new double[] { DFlockers.neighborhood / 1.5, DFlockers.neighborhood / 1.5 };
-		flockers = new DContinuous2D<>(DFlockers.neighborhood / 1.5, this);
+		flockers = new DContinuous2D<>((int)(DFlockers.neighborhood / 1.5), this);
 	}
 
 	@Override
